==> 00 - Aerodromes View.ipynb <==
import pandas as pd
import numpy as np

aerodromes = pd.read_csv('../aerodrome.csv', header=0, engine='python', encoding='utf-8')
full_aerodromes = pd.read_csv('../full_aerodrome.csv', engine='python', encoding='utf-8', error_bad_lines=False)

full_aerodromes_short = full_aerodromes[['Sigla', 'Latitude', 'Longitude']]
full_aerodromes_short.head()

aerodromes_coordinates = aerodromes.join(full_aerodromes_short.set_index('Sigla'), on='Sigla')
aerodromes_coordinates.head()

aerodromes_coordinates.to_csv('../aerodromes_coordinates.csv', header=1, index=False, encoding='utf-8')
