==> 00 - Organizing CSVs.ipynb <==
# Imports
import pandas as pd

# Loading the dataframe
df = pd.read_csv('../raw-data/Flights/flights.csv', header=0, engine='python', encoding='utf-8')
# df = pd.read_csv('raw-data/Voos_Brasil/2000/VRA_do_MES_2000_01.csv', header=0, delimiter=';', engine='python', encoding='iso-8859-1')

# Check the columns
df.columns

# df with company
company = pd.DataFrame()
company['Sigla'] = df['EMPRESA (SIGLA)']
company['Nome'] = df['EMPRESA (NOME)']
company['Nacionalidade'] = df['EMPRESA (NACIONALIDADE)']
company.drop_duplicates(inplace=True)

# df with aerodrome
aerodrome = pd.DataFrame()
aerodrome['Sigla'] = df['AEROPORTO DE ORIGEM (SIGLA)'].append(df['AEROPORTO DE DESTINO (SIGLA)'])
aerodrome['Nome'] = df['AEROPORTO DE ORIGEM (NOME)'].append(df['AEROPORTO DE DESTINO (NOME)'])
aerodrome['Pais'] = df['AEROPORTO DE ORIGEM (PAÍS)'].append(df['AEROPORTO DE DESTINO (PAÍS)'])
aerodrome['Continente'] = df['AEROPORTO DE ORIGEM (CONTINENTE)'].append(df['AEROPORTO DE DESTINO (CONTINENTE)'])
aerodrome.drop_duplicates(inplace=True)



final_df = pd.DataFrame()
final_df['Empresa'] = df['EMPRESA (SIGLA)']
final_df['Ano'] = df['ANO']
final_df['Mes'] = df['MÊS']
final_df['Aeroporto Origem'] = df['AEROPORTO DE ORIGEM (SIGLA)']
final_df['Aeroporto Origem Pais'] = df['AEROPORTO DE ORIGEM (PAÍS)']
final_df['Aeroporto Origem Continente'] = df['AEROPORTO DE ORIGEM (CONTINENTE)']
final_df['Aeroporto Destino'] = df['AEROPORTO DE DESTINO (SIGLA)']
final_df['Aeroporto Destino Pais'] = df['AEROPORTO DE DESTINO (PAÍS)']
final_df['Aeroporto Destino Continente'] = df['AEROPORTO DE DESTINO (CONTINENTE)']
final_df['Natureza'] = df['NATUREZA']
final_df['Grupo de Voo'] = df['GRUPO DE VOO']
final_df['Passageiros Pagos'] = df['PASSAGEIROS PAGOS']
final_df['Passageiros Gratis'] = df['PASSAGEIROS GRÁTIS']
final_df['Carga Paga'] = df['CARGA PAGA (KG)']
final_df['Carga Gratis'] = df['CARGA GRÁTIS (KG)']
final_df['Correio'] = df['CORREIO (KG)']
# final_df['ASK'] = df['ASK']
# final_df['RPK'] = df['RPK']
# final_df['ATK'] = df['ATK']
# final_df['RTK'] = df['RTK']
final_df['Distancia'] = df['DISTÂNCIA VOADA (KM)']
final_df['Decolagens'] = df['DECOLAGENS']
final_df['Assentos'] = df['ASSENTOS']
final_df['Payload'] = df['PAYLOAD']

del df

# Saving csvs
company.to_csv('company.csv', header=1, index=False, encoding='utf-8')
aerodrome.to_csv('aerodrome.csv', header=1, index=False, encoding='utf-8')
final_df.to_csv('flights.csv', header=1, index=False, encoding='utf-8')
