==> 02 - Data Prep.ipynb <==
# Imports
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# Loading the dataframe
flights = pd.read_csv('../flights.csv', header=0, engine='python', encoding='utf-8')

flights.shape

flights.isnull().sum()

# Nao se tem info do aeroporto de destino e nao tem como inferir um
# Como sao so 2 registros, vamos dropar
flights = flights[flights['Aeroporto Destino'].notnull()]
flights.isnull().sum()

# Há voos de um mesmo aeroporto para ele mesmo [?] (ex: 762108) com distancia percorrida 0
flights[flights['Passageiros Pagos'].isnull()].sort_values(['Ano', 'Mes'], ascending=False).head()

flights_without_passanger = flights['Passageiros Pagos'].isnull().groupby([flights['Empresa'], flights['Ano'], flights['Mes']]).sum().astype(int).reset_index(name='count')
flights_without_passanger[flights_without_passanger['count'] > 0].nlargest(20, 'count')

flights[flights['Assentos'].isnull()].sort_values(['Ano', 'Mes'], ascending=False).head()

flights_without_passanger = flights['Assentos'].isnull().groupby([flights['Empresa'], flights['Ano'], flights['Mes']]).sum().astype(int).reset_index(name='count')
flights_without_passanger[flights_without_passanger['count'] > 0].nlargest(20, 'count')

flights[flights['Ano'] == 2008].groupby(['Ano', 'Mes']).count()
# Estao faltando mtos dados GLO 2008 01  por exemplo, metade

# Esses são os problemas encontrados, para solucionar usaremos a media dos outros da mesma empresa
# no mesmo trecho dos outros meses, caso não haja, usar de outra companhia

# Inicialmente, tirar estatísticas básicas por empresa/origem/destino/natureza/grupo de voo
# depois completar os valores com os correspondentes

# Mas primeiro vamos verificar nossa hipótese

# Filling nulls with median
column_list = ['Passageiros Pagos', 'Passageiros Gratis', 'Carga Paga', 'Carga Gratis', 'Correio', 'Distancia', 'Decolagens', 'Assentos', 'Payload']
for i in column_list:
    flights[i] = flights.groupby(['Empresa', 'Aeroporto Origem', 'Aeroporto Destino', 'Natureza', 'Grupo de Voo'])[i].apply(lambda x: x.fillna(x.mean()))

column_list = ['Passageiros Pagos', 'Passageiros Gratis', 'Carga Paga', 'Carga Gratis', 'Correio', 'Distancia', 'Decolagens', 'Assentos', 'Payload']
for i in column_list:
    flights[i] = flights.groupby(['Aeroporto Origem', 'Aeroporto Destino', 'Natureza', 'Grupo de Voo'])[i].apply(lambda x: x.fillna(x.mean()))

column_list = ['Passageiros Pagos', 'Passageiros Gratis', 'Carga Paga', 'Carga Gratis', 'Correio', 'Distancia', 'Decolagens', 'Assentos', 'Payload']
for i in column_list:
    flights[i] = flights.groupby(['Natureza', 'Grupo de Voo'])[i].apply(lambda x: x.fillna(x.mean()))

flights = flights.drop(flights.index[flights['Aeroporto Origem Pais'].isnull()].tolist())
flights = flights.drop(flights.index[flights['Aeroporto Destino Pais'].isnull()].tolist())

# Criar coluna com total de passageiros, total de carga e porcentagem de passageiros
flights['Passageiros Totais'] = flights['Passageiros Pagos'] + flights['Passageiros Gratis']
flights['Carga Total'] = flights['Carga Paga'] + flights['Carga Gratis']
flights['Assentos Preenchidos'] = flights['Passageiros Totais']/flights['Assentos']

flights.head()

# Trocando por NaN
flights['Assentos Preenchidos'] = flights['Assentos Preenchidos'].replace([np.inf, -np.inf], np.nan)

flights['Assentos Preenchidos'] = flights.groupby(['Empresa', 'Aeroporto Origem', 'Aeroporto Destino', 'Natureza', 'Grupo de Voo'])['Assentos Preenchidos'].apply(lambda x: x.fillna(x.mean()))

flights['Assentos Preenchidos'] = flights.groupby(['Aeroporto Origem', 'Aeroporto Destino', 'Natureza', 'Grupo de Voo'])['Assentos Preenchidos'].apply(lambda x: x.fillna(x.mean()))

flights['Assentos Preenchidos'] = flights.groupby(['Natureza', 'Grupo de Voo'])['Assentos Preenchidos'].apply(lambda x: x.fillna(x.mean()))

flights = flights.drop(flights[flights['Assentos Preenchidos'].isnull()].index.tolist(), axis=0)

flights['Assentos Preenchidos'] = flights['Assentos Preenchidos'].apply(lambda x: 1 if x > 1 else x)

# Coluna que concatena o ano e mes, para dar alguma ordem
flights['Ano Mes'] = flights[['Ano', 'Mes']].apply(lambda x: int(str(x[0]) + str(x[1])) if x[1] >= 10 else int(str(x[0]) + '0' + str(x[1])), axis=1)

flights.to_csv('../flights_prepared.csv', header=1, index=False, encoding='utf-8')
