==> 03 - Building the Model.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

original_dataset = pd.read_csv('../flights_prepared.csv', header=0, engine='python', encoding='utf-8')

original_dataset.head()

dataset = original_dataset[['Empresa', 'Aeroporto Origem', 'Aeroporto Destino', 'Natureza', 'Ano Mes', 'Assentos Preenchidos']]
X = dataset.iloc[:, :-1]
y = dataset.iloc[:, 5]

# Encoding categorical data
# Encoding the Independent Variable
from sklearn.preprocessing import LabelEncoder, OneHotEncoder
le = LabelEncoder()
X_2 = X.apply(le.fit_transform)
X_2.head()

encoder = OneHotEncoder()

encoder.fit(X_2)

labels = encoder.transform(X_2).toarray()
labels.shape

from sklearn.cross_validation import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X_2, y, test_size = 0.2, random_state = 0)

from sklearn.ensemble import RandomForestRegressor
regressor = RandomForestRegressor(n_estimators = 10, random_state = 0)
regressor.fit(X_train, y_train)

y_pred = regressor.predict(X_test)

y_pred

from sklearn.metrics import mean_squared_error
from math import sqrt

rms = sqrt(mean_squared_error(y_test, y_pred))
rms
